:truck: Move: 불필요한 컴포넌트 제거 및 이동 (#110)
