+++ src/components/applicant/announcement_list/detail/AnnouncementContainer.tsx
