+++ src/components/applicant/announcement_list/detail/InqueryContainer.tsx
